build: remove some deps from git-branchless top-level package
