+++ git-branchless-opts/Cargo.toml
@@ -6,4 +6,5 @@ version = "0.7.0-pre.1"
 [dependencies]
 clap = { version = "4.0.23", features = ["derive"] }
 clap_mangen = "0.2.4"
+itertools = "0.10.5"
 lib = { package = "git-branchless-lib", version = "0.7.0-pre.1", path = "../git-branchless-lib" }
